% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Graph which correlates 2 variable and displays correlation coefficient on graph.}
\usage{
correlation(data = MER(), x = data$dep_delay, y = data$arr_delay,
  color1 = "forestgreen", color2 = "firebrick",
  color3 = "dodgerblue")
}
\arguments{
\item{data}{a data set}

\item{x}{a data frame.}

\item{y}{another data frame.}

\item{color1, 2, 3}{are different colors for each of the 3 variable which graph is grouped by.}
}
\value{
A graph displaying the correlation between two variables with correlation coefficient on graph.
}
\description{
Graph which correlates 2 variable and displays correlation coefficient on graph.
}
\examples{
#correlation()
}
