% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delaygraph.R
\name{Delaygraph}
\alias{Delaygraph}
\title{Graphing delays from NYC airpots}
\usage{
Delaygraph(data = MER(), color1 = "forestgreen",
  color2 = "firebrick", color3 = "dodgerblue")
}
\arguments{
\item{data}{a data set.}
}
\value{
A graph displaying delay time and distance for each of the three airports in NYC
}
\description{
Graphing delays from NYC airpots
}
\examples{
Delaygraph()
}
