% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MER.R
\name{MER}
\alias{MER}
\title{Merging two dataframes}
\usage{
MER(x = nycflights13::weather, y = nycflights13::flights)
}
\arguments{
\item{x}{a data frame.}

\item{y}{another data frame.}
}
\value{
A dataframe with \code{x} and \code{y} getting rid of duplicate columns.
}
\description{
Merging two dataframes
}
\examples{
#MER()
}
