% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speedgraph.R
\name{Speedgraph}
\alias{Speedgraph}
\title{Graphing delays from NYC airpots}
\usage{
Speedgraph(data = MER(), color1 = "forestgreen",
  color2 = "firebrick", color3 = "dodgerblue")
}
\arguments{
\item{data}{a data set.}
}
\value{
A graph displaying departure delay (min) and speed for each of the three airports in NYC
}
\description{
Graphing delays from NYC airpots
}
\examples{
Speedgraph()
}
